opnsense-management-api/reverse-proxy/api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
httpx==0.27.0
PyJWT==2.9.0
pydantic-settings==2.4.0